Moves test_pool_scheduling into its own file
